requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv
openai